+++ src/Client/ButtonModule.hpp
@@ -6,8 +6,6 @@ class ButtonModule : public Module
 {
     public:
         virtual void onClick() {}
-
-        bool shouldSave();
         
         virtual void onKeybindActivated(KeyState state);
         virtual ModuleNode* getNode();